Change filenames for figures
